requirements.txt
# Google Gemini API
google-generativeai>=0.3.0

# 웹 프레임워크
Flask>=3.0.0

# HTTP 요청
requests>=2.31.0

# 환경 변수 관리
python-dotenv>=1.0.0

# 타임존 처리
pytz>=2023.3

# (선택) 파일 감시 (watch 모드용)
# watchdog>=3.0.0
